Introduced CcdCaseSubmitter and ExceptionRecordProvider
