+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CreateCaseCallbackService.java
@@ -1,21 +1,13 @@
 package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 
 import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Maps;
-import feign.FeignException;
 import io.vavr.collection.Seq;
 import io.vavr.control.Try;
 import io.vavr.control.Validation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.InvalidCaseDataException;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.TransformationClient;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.request.ExceptionRecord;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.CaseCreationDetails;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.SuccessfulTransformationResponse;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.ServiceConfigItem;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.in.CcdCallbackRequest;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.CallbackException;
@@ -23,59 +15,41 @@ import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.CreateCas
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.ProcessResult;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.YesNoFieldValues;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.config.ServiceConfigProvider;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.payments.PaymentsPublishingException;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
 import static java.util.stream.Collectors.joining;
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.hasServiceNameInCaseTypeId;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.EventIdValidator.isCreateNewCaseEvent;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.AWAITING_PAYMENT_DCN_PROCESSING;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.CASE_REFERENCE;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.DISPLAY_WARNINGS;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.OCR_DATA_VALIDATION_WARNINGS;
 
 @Service
 public class CreateCaseCallbackService {
 
     private static final Logger log = LoggerFactory.getLogger(CreateCaseCallbackService.class);
 
-    private static final String EXCEPTION_RECORD_REFERENCE = "bulkScanCaseReference";
-
     private final CreateCaseValidator validator;
     private final ServiceConfigProvider serviceConfigProvider;
-    private final TransformationClient transformationClient;
-    private final AuthTokenGenerator s2sTokenGenerator;
-    private final PaymentsProcessor paymentsProcessor;
-    private final CoreCaseDataApi coreCaseDataApi;
     private final CcdApi ccdApi;
+    private final CcdCaseSubmitter ccdCaseSubmitter;
+    private final ExceptionRecordProvider exceptionRecordProvider;
 
     public CreateCaseCallbackService(
         CreateCaseValidator validator,
         ServiceConfigProvider serviceConfigProvider,
-        TransformationClient transformationClient,
-        AuthTokenGenerator s2sTokenGenerator,
-        PaymentsProcessor paymentsProcessor,
-        CoreCaseDataApi coreCaseDataApi,
-        CcdApi ccdApi
+        CcdApi ccdApi,
+        CcdCaseSubmitter ccdCaseSubmitter,
+        ExceptionRecordProvider exceptionRecordProvider
     ) {
         this.validator = validator;
         this.serviceConfigProvider = serviceConfigProvider;
-        this.transformationClient = transformationClient;
-        this.s2sTokenGenerator = s2sTokenGenerator;
-        this.paymentsProcessor = paymentsProcessor;
-        this.coreCaseDataApi = coreCaseDataApi;
         this.ccdApi = ccdApi;
+        this.ccdCaseSubmitter = ccdCaseSubmitter;
+        this.exceptionRecordProvider = exceptionRecordProvider;
     }
 
     /**
@@ -173,7 +147,7 @@ public class CreateCaseCallbackService {
         } else {
             List<Long> ids = ccdApi.getCaseRefsByBulkScanCaseReference(exceptionRecord.id, configItem.getService());
             if (ids.isEmpty()) {
-                return createNewCase(
+                return ccdCaseSubmitter.createNewCase(
                     exceptionRecord,
                     configItem,
                     ignoreWarnings,
@@ -183,7 +157,7 @@ public class CreateCaseCallbackService {
                 );
             } else if (ids.size() == 1) {
                 return new ProcessResult(
-                    prepareResultExceptionRecord(
+                    exceptionRecordProvider.prepareResultExceptionRecord(
                         exceptionRecordData.getData(),
                         ids.get(0)
                     )
@@ -201,178 +175,5 @@ public class CreateCaseCallbackService {
                 );
             }
         }
-    }
-
-    @SuppressWarnings({"squid:S2139", "unchecked"}) // squid for exception handle + logging
-    private ProcessResult createNewCase(
-        ExceptionRecord exceptionRecord,
-        ServiceConfigItem configItem,
-        boolean ignoreWarnings,
-        String idamToken,
-        String userId,
-        CaseDetails exceptionRecordData
-    ) {
-        log.info(
-            "Start creating new case for {} from exception record {}",
-            configItem.getService(),
-            exceptionRecord.id
-        );
-
-        try {
-            String s2sToken = s2sTokenGenerator.generate();
-
-            SuccessfulTransformationResponse transformationResponse = transformationClient.transformExceptionRecord(
-                configItem.getTransformationUrl(),
-                exceptionRecord,
-                s2sToken
-            );
-
-            if (!ignoreWarnings && !transformationResponse.warnings.isEmpty()) {
-                // do not log warnings
-                return new ProcessResult(transformationResponse.warnings, emptyList());
-            }
-
-            log.info(
-                "Successfully transformed exception record for {} from exception record {}",
-                configItem.getService(),
-                exceptionRecord.id
-            );
-
-            checkBulkScanReferenceIsSet(
-                (Map<String, ?>) transformationResponse.caseCreationDetails.caseData,
-                exceptionRecord.id
-            );
-
-            long newCaseId = createNewCaseInCcd(
-                idamToken,
-                s2sToken,
-                userId,
-                exceptionRecord,
-                transformationResponse.caseCreationDetails
-            );
-
-            log.info(
-                "Successfully created new case for {} with case ID {} from exception record {}",
-                configItem.getService(),
-                newCaseId,
-                exceptionRecord.id
-            );
-
-            paymentsProcessor.updatePayments(exceptionRecordData, newCaseId);
-
-            return new ProcessResult(
-                prepareResultExceptionRecord(exceptionRecordData.getData(), newCaseId)
-            );
-        } catch (InvalidCaseDataException exception) {
-            if (BAD_REQUEST.equals(exception.getStatus())) {
-                throw new CallbackException("Failed to transform exception record", exception);
-            } else {
-                return new ProcessResult(exception.getResponse().warnings, exception.getResponse().errors);
-            }
-        } catch (PaymentsPublishingException exception) {
-            log.error(
-                "Failed to send update to payment processor for {} exception record {}",
-                configItem.getService(),
-                exceptionRecord.id,
-                exception
-            );
-
-            throw new CallbackException("Payment references cannot be processed. Please try again later", exception);
-        } catch (Exception exception) {
-            // log happens individually to cover transformation/ccd cases
-            throw new CallbackException("Failed to create new case", exception);
-        }
-    }
-
-    @SuppressWarnings("squid:S2139") // exception handle + logging
-    private long createNewCaseInCcd(
-        String idamToken,
-        String s2sToken,
-        String userId,
-        ExceptionRecord exceptionRecord,
-        CaseCreationDetails caseCreationDetails
-    ) {
-        try {
-            StartEventResponse eventResponse = coreCaseDataApi.startForCaseworker(
-                idamToken,
-                s2sToken,
-                userId,
-                exceptionRecord.poBoxJurisdiction,
-                caseCreationDetails.caseTypeId,
-                // when onboarding remind services to not configure about to submit callback for this event
-                caseCreationDetails.eventId
-            );
-
-            return coreCaseDataApi.submitForCaseworker(
-                idamToken,
-                s2sToken,
-                userId,
-                exceptionRecord.poBoxJurisdiction,
-                caseCreationDetails.caseTypeId,
-                true,
-                CaseDataContent
-                    .builder()
-                    .caseReference(exceptionRecord.id)
-                    .data(caseCreationDetails.caseData)
-                    .event(Event
-                        .builder()
-                        .id(eventResponse.getEventId())
-                        .summary("Case created")
-                        .description("Case created from exception record ref " + exceptionRecord.id)
-                        .build()
-                    )
-                    .eventToken(eventResponse.getToken())
-                    .build()
-            ).getId();
-        } catch (FeignException exception) {
-            log.error(
-                "Failed to create new case for {} jurisdiction from exception record {}. Service response: {}",
-                exceptionRecord.poBoxJurisdiction,
-                exceptionRecord.id,
-                exception.contentUTF8(),
-                exception
-            );
-
-            throw exception;
-        } catch (Exception exception) {
-            log.error(
-                "Failed to create new case for {} jurisdiction from exception record {}",
-                exceptionRecord.poBoxJurisdiction,
-                exceptionRecord.id,
-                exception
-            );
-
-            throw exception;
-        }
-    }
-
-    private void checkBulkScanReferenceIsSet(Map<String, ?> caseData, String exceptionRecordId) {
-        if (!caseData.containsKey(EXCEPTION_RECORD_REFERENCE)) {
-            log.error(
-                "Transformation did not set '{}' with exception record id {}",
-                EXCEPTION_RECORD_REFERENCE,
-                exceptionRecordId
-            );
-        } else if (!caseData.get(EXCEPTION_RECORD_REFERENCE).equals(exceptionRecordId)) {
-            log.error(
-                "Transformation did not set exception record reference correctly. Actual: {}, expected: {}",
-                caseData.get(EXCEPTION_RECORD_REFERENCE),
-                exceptionRecordId
-            );
-        }
-    }
-
-    private Map<String, Object> prepareResultExceptionRecord(Map<String, Object> originalFields, Long caseReference) {
-        Map<String, Object> fieldsToUpdate =
-            ImmutableMap.<String, Object>builder()
-                .put(CASE_REFERENCE, Long.toString(caseReference))
-                .put(DISPLAY_WARNINGS, YesNoFieldValues.NO)
-                .put(OCR_DATA_VALIDATION_WARNINGS, emptyList())
-                .build();
-
-        return ImmutableMap.<String, Object>builder()
-            .putAll(Maps.difference(originalFields, fieldsToUpdate).entriesOnlyOnLeft())
-            .putAll(fieldsToUpdate)
-            .build();
     }
 }
