+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CreateCaseCallbackServiceTest.java
@@ -7,17 +7,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.client.HttpClientErrorException;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.CaseTransformationException;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.InvalidCaseDataException;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.TransformationClient;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.request.DocumentType;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.request.ExceptionRecord;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.CaseCreationDetails;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.SuccessfulTransformationResponse;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.client.transformation.model.response.TransformationErrorResponse;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.ServiceConfigItem;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.in.CcdCallbackRequest;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.CallbackException;
@@ -28,35 +20,23 @@ import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.YesNoFi
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.config.ServiceConfigProvider;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.config.ServiceNotConfiguredException;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.payments.PaymentsPublishingException;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
 import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
-import static java.util.Collections.emptyMap;
-import static java.util.Collections.singletonList;
-import static java.util.UUID.randomUUID;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowableOfType;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.EventIdValidator.EVENT_ID_CREATE_NEW_CASE;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.CASE_REFERENCE;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.EXCEPTION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.NEW_APPLICATION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE;
@@ -89,6 +69,12 @@ class CreateCaseCallbackServiceTest {
     @Mock
     private PaymentsProcessor paymentsProcessor;
 
+    @Mock
+    private CcdCaseSubmitter ccdCaseSubmitter;
+
+    @Mock
+    private ExceptionRecordProvider exceptionRecordProvider;
+
     private CreateCaseCallbackService service;
 
     @BeforeEach
@@ -96,11 +82,9 @@ class CreateCaseCallbackServiceTest {
         service = new CreateCaseCallbackService(
             VALIDATOR,
             serviceConfigProvider,
-            transformationClient,
-            s2sTokenGenerator,
-            paymentsProcessor,
-            coreCaseDataApi,
-            ccdApi
+            ccdApi,
+            ccdCaseSubmitter,
+            exceptionRecordProvider
         );
     }
 
@@ -117,7 +101,9 @@ class CreateCaseCallbackServiceTest {
 
         assertThat(callbackException.getCause()).isNull();
         assertThat(callbackException).hasMessage("The some event event is not supported. Please contact service team");
+
         verify(serviceConfigProvider, never()).getConfig(anyString());
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -137,7 +123,9 @@ class CreateCaseCallbackServiceTest {
 
         assertThat(callbackException.getCause()).isNull();
         assertThat(callbackException).hasMessage("No case type ID supplied");
+
         verify(serviceConfigProvider, never()).getConfig(anyString());
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -158,7 +146,9 @@ class CreateCaseCallbackServiceTest {
         // then
         assertThat(callbackException.getCause()).isNull();
         assertThat(callbackException).hasMessage("Case type ID () has invalid format");
+
         verify(serviceConfigProvider, never()).getConfig(anyString());
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -180,6 +170,8 @@ class CreateCaseCallbackServiceTest {
         // then
         assertThat(callbackException.getCause()).isNull();
         assertThat(callbackException).hasMessage("oh no");
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -201,6 +193,8 @@ class CreateCaseCallbackServiceTest {
         // then
         assertThat(callbackException.getCause()).isNull();
         assertThat(callbackException).hasMessage("Transformation URL is not configured");
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -233,6 +227,8 @@ class CreateCaseCallbackServiceTest {
                 NEW_APPLICATION.name()
             )
         );
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -259,54 +255,8 @@ class CreateCaseCallbackServiceTest {
                 SUPPLEMENTARY_EVIDENCE.name()
             )
         );
-    }
 
-    @Test
-    void should_throw_InvalidCaseDataException_when_transformation_client_returns_422()
-        throws IOException, CaseTransformationException {
-        // given
-        when(s2sTokenGenerator.generate()).thenReturn(randomUUID().toString());
-        setUpTransformationUrl();
-        InvalidCaseDataException exception = new InvalidCaseDataException(
-            new HttpClientErrorException(HttpStatus.UNPROCESSABLE_ENTITY),
-            new TransformationErrorResponse(singletonList("error"), singletonList("warning"))
-        );
-        doThrow(exception)
-            .when(transformationClient)
-            .transformExceptionRecord(anyString(), any(ExceptionRecord.class), anyString());
-
-        // when
-        ProcessResult result = service.process(new CcdCallbackRequest(
-            EVENT_ID_CREATE_NEW_CASE,
-            caseDetails(basicCaseData()),
-            true
-        ), IDAM_TOKEN, USER_ID);
-
-        // then
-        assertThat(result.getExceptionRecordData()).isEmpty();
-        assertThat(result.getWarnings()).containsOnly("warning");
-        assertThat(result.getErrors()).containsOnly("error");
-    }
-
-    @Test
-    void should_return_service_case_when_it_already_exists_in_ccd_for_a_given_exception_record() throws Exception {
-        // given
-        setUpTransformationUrl();
-
-        when(ccdApi.getCaseRefsByBulkScanCaseReference(Long.toString(CASE_ID), null))
-            .thenReturn(singletonList(345L));
-
-        // when
-        ProcessResult result = service.process(new CcdCallbackRequest(
-            EVENT_ID_CREATE_NEW_CASE,
-            caseDetails(basicCaseData()),
-            true
-        ), IDAM_TOKEN, USER_ID);
-
-        // then
-        assertThat(result.getExceptionRecordData().get(CASE_REFERENCE)).isEqualTo("345");
-        assertThat(result.getWarnings().isEmpty()).isTrue();
-        assertThat(result.getErrors().isEmpty()).isTrue();
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -332,6 +282,8 @@ class CreateCaseCallbackServiceTest {
         assertThat(result.getErrors()).containsOnly(
             "Multiple cases (345, 456) found for the given bulk scan case reference: 123"
         );
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -354,6 +306,8 @@ class CreateCaseCallbackServiceTest {
         // then
         assertThat(result.getWarnings()).isEmpty();
         assertThat(result.getErrors()).containsOnly("Missing journeyClassification");
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -376,6 +330,8 @@ class CreateCaseCallbackServiceTest {
         assertThat(result.getErrors()).containsOnly(
             "Invalid journeyClassification. Error: No enum constant " + Classification.class.getName() + ".EXCEPTIONS"
         );
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -422,6 +378,8 @@ class CreateCaseCallbackServiceTest {
         assertThat(result.getErrors()).containsOnly(
             "Invalid scannedDocuments format. Error: No enum constant " + DocumentType.class.getName() + ".OTHERS"
         );
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -477,6 +435,8 @@ class CreateCaseCallbackServiceTest {
             .element(0)
             .asString()
             .matches(match);
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -502,152 +462,8 @@ class CreateCaseCallbackServiceTest {
             .element(0)
             .asString()
             .matches(match);
-    }
 
-    @Test
-    void should_call_payments_processor_when_case_has_payments() throws Exception {
-        // given
-        setUpTransformationUrl();
-
-        given(transformationClient.transformExceptionRecord(any(), any(), any()))
-            .willReturn(
-                new SuccessfulTransformationResponse(
-                    new CaseCreationDetails(
-                        "some_case_type",
-                        "some_event_id",
-                        emptyMap()
-                    ),
-                    emptyList()
-                )
-            );
-
-        StartEventResponse startCcdEventResp = mock(StartEventResponse.class);
-
-        given(coreCaseDataApi.startForCaseworker(any(), any(), any(), any(), any(), any()))
-            .willReturn(startCcdEventResp);
-
-        Long newCaseId = 123L;
-        CaseDetails newCaseDetails = mock(CaseDetails.class);
-        doReturn(newCaseId).when(newCaseDetails).getId();
-
-        given(coreCaseDataApi.submitForCaseworker(any(), any(), any(), any(), any(), anyBoolean(), any()))
-            .willReturn(newCaseDetails);
-
-        Map<String, Object> data = basicCaseData();
-        data.put(ExceptionRecordFields.CONTAINS_PAYMENTS, YesNoFieldValues.YES); // has payments!
-
-        CaseDetails caseDetails = caseDetails(data);
-
-        // when
-        ProcessResult result =
-            service
-                .process(
-                    new CcdCallbackRequest(EVENT_ID_CREATE_NEW_CASE, caseDetails, true),
-                    IDAM_TOKEN,
-                    USER_ID
-                );
-
-        assertThat(result.getErrors()).isEmpty();
-        assertThat(result.getWarnings()).isEmpty();
-
-        verify(paymentsProcessor).updatePayments(caseDetails, CASE_ID);
-    }
-
-    @Test
-    void should_call_payments_processor_when_case_has_no_payments() throws Exception {
-        // given
-        setUpTransformationUrl();
-
-        given(transformationClient.transformExceptionRecord(any(), any(), any()))
-            .willReturn(
-                new SuccessfulTransformationResponse(
-                    new CaseCreationDetails(
-                        "some_case_type",
-                        "some_event_id",
-                        emptyMap()
-                    ),
-                    emptyList()
-                )
-            );
-
-        StartEventResponse startCcdEventResp = mock(StartEventResponse.class);
-
-        given(coreCaseDataApi.startForCaseworker(any(), any(), any(), any(), any(), any()))
-            .willReturn(startCcdEventResp);
-
-        CaseDetails newCaseDetails = mock(CaseDetails.class);
-        doReturn(123L).when(newCaseDetails).getId();
-
-        given(coreCaseDataApi.submitForCaseworker(any(), any(), any(), any(), any(), anyBoolean(), any()))
-            .willReturn(newCaseDetails);
-
-        Map<String, Object> data = basicCaseData();
-        data.put(ExceptionRecordFields.CONTAINS_PAYMENTS, YesNoFieldValues.NO); // no payments!
-
-        CaseDetails caseDetails = caseDetails(data);
-
-        // when
-        ProcessResult result =
-            service
-                .process(
-                    new CcdCallbackRequest(EVENT_ID_CREATE_NEW_CASE, caseDetails, true),
-                    IDAM_TOKEN,
-                    USER_ID
-                );
-
-        // then
-        assertThat(result.getErrors()).isEmpty();
-        assertThat(result.getWarnings()).isEmpty();
-
-        verify(paymentsProcessor).updatePayments(caseDetails, CASE_ID);
-    }
-
-    @Test
-    void should_return_error_if_publishing_payment_command_failed() throws Exception {
-        // given
-        setUpTransformationUrl();
-
-        given(transformationClient.transformExceptionRecord(any(), any(), any()))
-            .willReturn(
-                new SuccessfulTransformationResponse(
-                    new CaseCreationDetails("some_case_type", "some_event_id", emptyMap()),
-                    emptyList()
-                )
-            );
-
-        StartEventResponse startCcdEventResp = mock(StartEventResponse.class);
-
-        given(coreCaseDataApi.startForCaseworker(any(), any(), any(), any(), any(), any()))
-            .willReturn(startCcdEventResp);
-
-        Long newCaseId = 123L;
-        CaseDetails newCaseDetails = mock(CaseDetails.class);
-        doReturn(newCaseId).when(newCaseDetails).getId();
-
-        given(coreCaseDataApi.submitForCaseworker(any(), any(), any(), any(), any(), anyBoolean(), any()))
-            .willReturn(newCaseDetails);
-
-        CaseDetails caseDetails = caseDetails(basicCaseData());
-
-        PaymentsPublishingException paymentException = new PaymentsPublishingException("Error message", null);
-        doThrow(paymentException)
-            .when(paymentsProcessor)
-            .updatePayments(any(), anyLong());
-
-        // when
-        CallbackException exception = catchThrowableOfType(() ->
-            service
-                .process(
-                    new CcdCallbackRequest(EVENT_ID_CREATE_NEW_CASE, caseDetails, true),
-                    IDAM_TOKEN,
-                    USER_ID
-                ),
-            CallbackException.class
-        );
-
-        // then
-        assertThat(exception.getMessage()).isEqualTo("Payment references cannot be processed. Please try again later");
-        assertThat(exception.getCause()).isInstanceOf(PaymentsPublishingException.class);
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     @Test
@@ -670,6 +486,8 @@ class CreateCaseCallbackServiceTest {
         // then
         assertThat(result.getWarnings())
             .containsExactly("Payments for this Exception Record have not been processed yet");
+
+        verify(exceptionRecordProvider, never()).prepareResultExceptionRecord(anyMap(), anyLong());
     }
 
     private void setUpTransformationUrl() {
